requirements.txt
aiohttp
python-socketio[asyncio_client]
redis
python-socketio